Fix small circular reference
